fix: fix some boundary case when log_time_monitor_pid may not be there any more

+++ benchmark-scripts/benchmark.sh
@@ -321,12 +321,17 @@ do
 
  sleep 2
 
- pkill -P $log_time_monitor_pid
- while ps -p $log_time_monitor_pid > /dev/null
- do
-		echo "$log_time_monitor_pid is still running"
-		sleep 1
- done
+ # before kill the process check if it is already gone
+ if ps -p $log_time_monitor_pid > /dev/null
+ then
+    pkill -P $log_time_monitor_pid
+    while ps -p $log_time_monitor_pid > /dev/null
+    do
+		  echo "$log_time_monitor_pid is still running"
+		  sleep 1
+    done
+ fi
+ 
  ./stop_server.sh
  sleep 5
